Untangle cyclic deps configuration <> secrets
